# Conflicts:
#	tools/train-intent/README.md
#	tools/train-intent/config.yaml
#	tools/train-intent/queue_bakeoff.sh
#	tools/train-intent/seeds/intents.yaml
#	tools/train-intent/src/build_dataset.py
#	tools/train-intent/src/eval_golden.py
#	tools/train-intent/src/gen_teacher.py
#	tools/train-intent/src/stt_noise.py
#	tools/train-intent/src/train_qlora.py

+++ tools/train/src/eval_checkpoint.py
@@ -67,7 +67,8 @@ def main() -> None:
     cfg = apply_common(cfg, args)
 
     data_dir = abs_path(cfg, "data_dir")
-    model = WhisperForConditionalGeneration.from_pretrained(args.model)
+    model = WhisperForConditionalGeneration.from_pretrained(
+        args.model, torch_dtype=torch.float32)
     if args.processor == "medium":
         from transformers import WhisperProcessor
         processor = WhisperProcessor.from_pretrained(
